move hist test
